Fix missing float call

+++ birdclef/models/classifier_nn/datasets.py
@@ -22,7 +22,7 @@ class ToFloatTensor:
         self.device = device
 
     def __call__(self, sample):
-        z = [torch.from_numpy(z) for z in sample]
+        z = [torch.from_numpy(z).float() for z in sample]
         if self.device is not None:
             z = [z.to(self.device) for z in z]
         return tuple(z)
